refactor(project): update pages to use DTO pattern (Phase 22c)
- Refactor Project creation and edit pages to use CreateProjectRequest and UpdateProjectRequest
- Update imports to use ProjectFiles from the domain layer
- Remove .id! assertions and redundant checks in Project UI pages
- Map UI state directly to required fields in new mutation DTO signatures
